Default all events to use sendBeacon
https://developers.google.com/analytics/devguides/collection/analyticsjs
/field-reference#transport

http://www.w3.org/TR/beacon/#sec-sendBeacon-method

`sendBeacon` is a way to send small packets of data asynchronously,
data may be sent with a lower priority than other network traffic but
that data will be sent even if the page is closed or navigated away
from.

The browser should give priority to the network things the user is
doing, while our analytics should also be more complete.

Universal Analytics automatically falls back to the usual xhr or image
transport methods if sendBeacon is not supported.